added common.h/ split joint_async_copy/async_work_group_copy tests.
